Moved errros to new file, removed some  unused imports
