+++ src/utilities/errors.ts
@@ -1,0 +1,11 @@
+class HttpError extends Error {
+    status: number;
+    statusText: string;
+    constructor(message: string, status: number, statusText: string) {
+      super(message);
+      this.status = status;
+      this.statusText = statusText;
+    }
+  }
+
+  export { HttpError }
